requirements.txt
flask
flask-sqlalchemy
flask-restful
sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
# Flask-Rest-JSONAPI
# flask-jwt-simple
